Date double value format fix. Cannot convert to Long because it's not milliseconds

+++ src/commonMain/kotlin/com/configcat/log/LogHelper.kt
@@ -63,8 +63,7 @@ internal object LogHelper {
         }
 
         return if (isDate) {
-            // TODO E format, this is not long
-            val comparisonValueString = comparisonValue.toLong().toString()
+            val comparisonValueString = comparisonValue.toDouble().toString()
             "'$comparisonValueString' (${comparisonValue.toDateTimeUTCString()} UTC)"
         } else {
             var comparisonValueString = comparisonValue.toString()
